refactor: InputUtil -> InputView
